requirements.txt
python-dotenv
PyPDF2
sentence-transformers
faiss-cpu
azure-ai-inference
mistralai